+++ packages/expo-audio/ios/AudioModule.swift
@@ -200,6 +200,25 @@ public class AudioModule: Module {
         }
       }
 
+      Function("setActiveForLockScreen") { (player: AudioPlayer, active: Bool, metadata: Metadata?) in
+        player.setActiveForLockScreen(active, metadata: metadata)
+      }
+
+      Function("updateLockScreenMetadata") { (player: AudioPlayer, metadata: Metadata?) in
+        if player.isActiveForLockScreen {
+          player.metadata = metadata
+          MediaController.shared.updateNowPlayingInfo(for: player)
+        }
+      }
+
+      Function("clearLockScreenControls") { (player: AudioPlayer) in
+        if player.isActiveForLockScreen {
+          player.metadata = nil
+          player.isActiveForLockScreen = false
+          MediaController.shared.setActivePlayer(nil)
+        }
+      }
+
       AsyncFunction("seekTo") { (player: AudioPlayer, seconds: Double, toleranceMillisBefore: Double?, toleranceMillisAfter: Double?) in
         await player.seekTo(
           seconds: seconds,
@@ -468,7 +487,7 @@ public class AudioModule: Module {
     try AudioUtils.validateAudioMode(mode: mode)
     let session = AVAudioSession.sharedInstance()
     var category: AVAudioSession.Category = session.category
-
+    
     self.shouldPlayInBackground = mode.shouldPlayInBackground
     self.interruptionMode = mode.interruptionMode
     self.allowsRecording = mode.allowsRecording
@@ -517,6 +536,7 @@ public class AudioModule: Module {
       sessionOptions = categoryOptions
     }
 
+<<<<<<< HEAD
     try session.setCategory(category, options: sessionOptions)
   }
 
@@ -539,6 +559,17 @@ public class AudioModule: Module {
         }
       }
     }
+||||||| parent of dbe8b959ef ([ios][audio] Support lock screen controls)
+    try session.setCategory(category, options: sessionOptions)
+    try session.setActive(true, options: [.notifyOthersOnDeactivation])
+=======
+    if sessionOptions.isEmpty {
+      try session.setCategory(category, mode: .default)
+    } else {
+      try session.setCategory(category, options: sessionOptions)
+    }
+    try session.setActive(true, options: [.notifyOthersOnDeactivation])
+>>>>>>> dbe8b959ef ([ios][audio] Support lock screen controls)
   }
 
   private func checkPermissions() throws {